requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
azure-core==1.33.0
    # via azure-storage-blob
azure-storage-blob==12.25.1
    # via -r requirements.in
bump2version==1.0.1
    # via -r requirements.in
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
cryptography==44.0.2
    # via azure-storage-blob
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi[standard]==0.115.12
    # via -r requirements.in
fastapi-cli[standard]==0.0.7
    # via fastapi
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2025.3.2
    # via
    #   huggingface-hub
    #   torch
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via fastapi
huggingface-hub==0.30.2
    # via
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
isodate==0.7.2
    # via azure-storage-blob
jinja2==3.1.6
    # via
    #   fastapi
    #   torch
joblib==1.4.2
    # via scikit-learn
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==2.2.4
    # via
    #   pandas
    #   scikit-learn
    #   scipy
    #   transformers
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.2
    # via torch
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
packaging==24.2
    # via
    #   huggingface-hub
    #   transformers
pandas==2.2.3
    # via -r requirements.in
pycparser==2.22
    # via cffi
pydantic==2.11.3
    # via fastapi
pydantic-core==2.33.1
    # via pydantic
pygments==2.19.1
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   transformers
    #   uvicorn
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   azure-core
    #   huggingface-hub
    #   transformers
rich==14.0.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.14.1
    # via fastapi-cli
safetensors==0.5.3
    # via transformers
scikit-learn==1.6.1
    # via -r requirements.in
scipy==1.15.2
    # via scikit-learn
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   azure-core
    #   python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.46.1
    # via fastapi
sympy==1.13.1
    # via torch
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.21.1
    # via transformers
torch==2.6.0
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   transformers
transformers==4.51.2
    # via -r requirements.in
triton==3.2.0
    # via torch
typer==0.15.2
    # via fastapi-cli
typing-extensions==4.13.2
    # via
    #   anyio
    #   azure-core
    #   azure-storage-blob
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   torch
    #   typer
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via requests
uvicorn[standard]==0.34.0
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
